Get synced with NWCHemEx Integrals

+++ Integrals/FourCTensorBuilder.cpp
@@ -12,8 +12,8 @@ std::vector<typename IntegralTensorBuilder<4>::TensorType> kernel(const LibChemi
     libint_type libints(0,molecule,basissets[0],basissets[1],basissets[2],basissets[3]);
     Integrals::FourCenterIntegral *Ints = &libints;
     IntegralTensorBuilder<4>::TensorType init_tensor(std::array<long int,4>
-                                                     {basissets[0].size(),basissets[1].size(),
-                                                      basissets[2].size(),basissets[3].size()});
+                                                     {static_cast<long>(basissets[0].size()),static_cast<long>(basissets[1].size()),
+                                                      static_cast<long>(basissets[2].size()),static_cast<long>(basissets[3].size())});
     init_tensor.setZero();
     std::vector<IntegralTensorBuilder<4>::TensorType> rv(Ints->n_components(),init_tensor);
